+++ test/Api.Test/Billing/Controllers/ProviderBillingControllerTests.cs
@@ -5,18 +5,14 @@ using Bit.Core;
 using Bit.Core.AdminConsole.Entities.Provider;
 using Bit.Core.AdminConsole.Enums.Provider;
 using Bit.Core.AdminConsole.Repositories;
-using Bit.Core.Billing.Constants;
 using Bit.Core.Billing.Enums;
 using Bit.Core.Billing.Models;
 using Bit.Core.Billing.Services;
 using Bit.Core.Context;
-using Bit.Core.Enums;
-using Bit.Core.Models.Api;
 using Bit.Core.Services;
 using Bit.Core.Utilities;
 using Bit.Test.Common.AutoFixture;
 using Bit.Test.Common.AutoFixture.Attributes;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using NSubstitute;
 using NSubstitute.ReturnsExtensions;
@@ -31,7 +27,74 @@ namespace Bit.Api.Test.Billing.Controllers;
 [SutProviderCustomize]
 public class ProviderBillingControllerTests
 {
-    #region GetInvoicesAsync
+    #region GetInvoicesAsync & TryGetBillableProviderForAdminOperations
+
+    [Theory, BitAutoData]
+    public async Task GetInvoicesAsync_FFDisabled_NotFound(
+        Guid providerId,
+        SutProvider<ProviderBillingController> sutProvider)
+    {
+        sutProvider.GetDependency<IFeatureService>().IsEnabled(FeatureFlagKeys.EnableConsolidatedBilling)
+            .Returns(false);
+
+        var result = await sutProvider.Sut.GetInvoicesAsync(providerId);
+
+        AssertNotFound(result);
+    }
+
+    [Theory, BitAutoData]
+    public async Task GetInvoicesAsync_NullProvider_NotFound(
+        Guid providerId,
+        SutProvider<ProviderBillingController> sutProvider)
+    {
+        sutProvider.GetDependency<IFeatureService>().IsEnabled(FeatureFlagKeys.EnableConsolidatedBilling)
+            .Returns(true);
+
+        sutProvider.GetDependency<IProviderRepository>().GetByIdAsync(providerId).ReturnsNull();
+
+        var result = await sutProvider.Sut.GetInvoicesAsync(providerId);
+
+        AssertNotFound(result);
+    }
+
+    [Theory, BitAutoData]
+    public async Task GetInvoicesAsync_NotProviderUser_Unauthorized(
+        Provider provider,
+        SutProvider<ProviderBillingController> sutProvider)
+    {
+        sutProvider.GetDependency<IFeatureService>().IsEnabled(FeatureFlagKeys.EnableConsolidatedBilling)
+            .Returns(true);
+
+        sutProvider.GetDependency<IProviderRepository>().GetByIdAsync(provider.Id).Returns(provider);
+
+        sutProvider.GetDependency<ICurrentContext>().ProviderProviderAdmin(provider.Id)
+            .Returns(false);
+
+        var result = await sutProvider.Sut.GetInvoicesAsync(provider.Id);
+
+        AssertUnauthorized(result);
+    }
+
+    [Theory, BitAutoData]
+    public async Task GetInvoicesAsync_ProviderNotBillable_Unauthorized(
+        Provider provider,
+        SutProvider<ProviderBillingController> sutProvider)
+    {
+        sutProvider.GetDependency<IFeatureService>().IsEnabled(FeatureFlagKeys.EnableConsolidatedBilling)
+            .Returns(true);
+
+        provider.Type = ProviderType.Reseller;
+        provider.Status = ProviderStatusType.Created;
+
+        sutProvider.GetDependency<IProviderRepository>().GetByIdAsync(provider.Id).Returns(provider);
+
+        sutProvider.GetDependency<ICurrentContext>().ProviderProviderAdmin(provider.Id)
+            .Returns(true);
+
+        var result = await sutProvider.Sut.GetInvoicesAsync(provider.Id);
+
+        AssertUnauthorized(result);
+    }
 
     [Theory, BitAutoData]
     public async Task GetInvoices_Ok(
@@ -131,158 +194,6 @@ public class ProviderBillingControllerTests
 
         Assert.Equal("text/csv", response.ContentType);
         Assert.Equal(reportContent, response.FileContents);
-    }
-
-    #endregion
-
-    #region GetPaymentInformationAsync & TryGetBillableProviderForAdminOperation
-
-    [Theory, BitAutoData]
-    public async Task GetPaymentInformationAsync_FFDisabled_NotFound(
-        Guid providerId,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        sutProvider.GetDependency<IFeatureService>().IsEnabled(FeatureFlagKeys.EnableConsolidatedBilling)
-            .Returns(false);
-
-        var result = await sutProvider.Sut.GetPaymentInformationAsync(providerId);
-
-        AssertNotFound(result);
-    }
-
-    [Theory, BitAutoData]
-    public async Task GetPaymentInformationAsync_NullProvider_NotFound(
-        Guid providerId,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        sutProvider.GetDependency<IFeatureService>().IsEnabled(FeatureFlagKeys.EnableConsolidatedBilling)
-            .Returns(true);
-
-        sutProvider.GetDependency<IProviderRepository>().GetByIdAsync(providerId).ReturnsNull();
-
-        var result = await sutProvider.Sut.GetPaymentInformationAsync(providerId);
-
-        AssertNotFound(result);
-    }
-
-    [Theory, BitAutoData]
-    public async Task GetPaymentInformationAsync_NotProviderUser_Unauthorized(
-        Provider provider,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        sutProvider.GetDependency<IFeatureService>().IsEnabled(FeatureFlagKeys.EnableConsolidatedBilling)
-            .Returns(true);
-
-        sutProvider.GetDependency<IProviderRepository>().GetByIdAsync(provider.Id).Returns(provider);
-
-        sutProvider.GetDependency<ICurrentContext>().ProviderProviderAdmin(provider.Id)
-            .Returns(false);
-
-        var result = await sutProvider.Sut.GetPaymentInformationAsync(provider.Id);
-
-        AssertUnauthorized(result);
-    }
-
-    [Theory, BitAutoData]
-    public async Task GetPaymentInformationAsync_ProviderNotBillable_Unauthorized(
-        Provider provider,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        sutProvider.GetDependency<IFeatureService>().IsEnabled(FeatureFlagKeys.EnableConsolidatedBilling)
-            .Returns(true);
-
-        provider.Type = ProviderType.Reseller;
-        provider.Status = ProviderStatusType.Created;
-
-        sutProvider.GetDependency<IProviderRepository>().GetByIdAsync(provider.Id).Returns(provider);
-
-        sutProvider.GetDependency<ICurrentContext>().ProviderProviderAdmin(provider.Id)
-            .Returns(true);
-
-        var result = await sutProvider.Sut.GetPaymentInformationAsync(provider.Id);
-
-        AssertUnauthorized(result);
-    }
-
-    [Theory, BitAutoData]
-    public async Task GetPaymentInformation_PaymentInformationNull_NotFound(
-        Provider provider,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        ConfigureStableAdminInputs(provider, sutProvider);
-
-        sutProvider.GetDependency<ISubscriberService>().GetPaymentInformation(provider).ReturnsNull();
-
-        var result = await sutProvider.Sut.GetPaymentInformationAsync(provider.Id);
-
-        Assert.IsType<NotFound>(result);
-    }
-
-    [Theory, BitAutoData]
-    public async Task GetPaymentInformation_Ok(
-        Provider provider,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        ConfigureStableAdminInputs(provider, sutProvider);
-
-        var maskedPaymentMethod = new MaskedPaymentMethodDTO(PaymentMethodType.Card, "VISA *1234", false);
-
-        var taxInformation =
-            new TaxInformationDTO("US", "12345", "123456789", "123 Example St.", null, "Example Town", "NY");
-
-        sutProvider.GetDependency<ISubscriberService>().GetPaymentInformation(provider).Returns(new PaymentInformationDTO(
-            100,
-            maskedPaymentMethod,
-            taxInformation));
-
-        var result = await sutProvider.Sut.GetPaymentInformationAsync(provider.Id);
-
-        Assert.IsType<Ok<PaymentInformationResponse>>(result);
-
-        var response = ((Ok<PaymentInformationResponse>)result).Value;
-
-        Assert.Equal(100, response.AccountCredit);
-        Assert.Equal(maskedPaymentMethod.Description, response.PaymentMethod.Description);
-        Assert.Equal(taxInformation.TaxId, response.TaxInformation.TaxId);
-    }
-
-    #endregion
-
-    #region GetPaymentMethodAsync
-
-    [Theory, BitAutoData]
-    public async Task GetPaymentMethod_PaymentMethodNull_NotFound(
-        Provider provider,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        ConfigureStableAdminInputs(provider, sutProvider);
-
-        sutProvider.GetDependency<ISubscriberService>().GetPaymentMethod(provider).ReturnsNull();
-
-        var result = await sutProvider.Sut.GetPaymentMethodAsync(provider.Id);
-
-        Assert.IsType<NotFound>(result);
-    }
-
-    [Theory, BitAutoData]
-    public async Task GetPaymentMethod_Ok(
-        Provider provider,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        ConfigureStableAdminInputs(provider, sutProvider);
-
-        sutProvider.GetDependency<ISubscriberService>().GetPaymentMethod(provider).Returns(new MaskedPaymentMethodDTO(
-            PaymentMethodType.Card, "Description", false));
-
-        var result = await sutProvider.Sut.GetPaymentMethodAsync(provider.Id);
-
-        Assert.IsType<Ok<MaskedPaymentMethodResponse>>(result);
-
-        var response = ((Ok<MaskedPaymentMethodResponse>)result).Value;
-
-        Assert.Equal(PaymentMethodType.Card, response.Type);
-        Assert.Equal("Description", response.Description);
-        Assert.False(response.NeedsVerification);
     }
 
     #endregion
@@ -454,78 +365,6 @@ public class ProviderBillingControllerTests
 
     #endregion
 
-    #region GetTaxInformationAsync
-
-    [Theory, BitAutoData]
-    public async Task GetTaxInformation_TaxInformationNull_NotFound(
-        Provider provider,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        ConfigureStableAdminInputs(provider, sutProvider);
-
-        sutProvider.GetDependency<ISubscriberService>().GetTaxInformation(provider).ReturnsNull();
-
-        var result = await sutProvider.Sut.GetTaxInformationAsync(provider.Id);
-
-        Assert.IsType<NotFound>(result);
-    }
-
-    [Theory, BitAutoData]
-    public async Task GetTaxInformation_Ok(
-        Provider provider,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        ConfigureStableAdminInputs(provider, sutProvider);
-
-        sutProvider.GetDependency<ISubscriberService>().GetTaxInformation(provider).Returns(new TaxInformationDTO(
-            "US",
-            "12345",
-            "123456789",
-            "123 Example St.",
-            null,
-            "Example Town",
-            "NY"));
-
-        var result = await sutProvider.Sut.GetTaxInformationAsync(provider.Id);
-
-        Assert.IsType<Ok<TaxInformationResponse>>(result);
-
-        var response = ((Ok<TaxInformationResponse>)result).Value;
-
-        Assert.Equal("US", response.Country);
-        Assert.Equal("12345", response.PostalCode);
-        Assert.Equal("123456789", response.TaxId);
-        Assert.Equal("123 Example St.", response.Line1);
-        Assert.Null(response.Line2);
-        Assert.Equal("Example Town", response.City);
-        Assert.Equal("NY", response.State);
-    }
-
-    #endregion
-
-    #region UpdatePaymentMethodAsync
-
-    [Theory, BitAutoData]
-    public async Task UpdatePaymentMethod_Ok(
-        Provider provider,
-        TokenizedPaymentMethodRequestBody requestBody,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        ConfigureStableAdminInputs(provider, sutProvider);
-
-        await sutProvider.Sut.UpdatePaymentMethodAsync(provider.Id, requestBody);
-
-        await sutProvider.GetDependency<ISubscriberService>().Received(1).UpdatePaymentMethod(
-            provider, Arg.Is<TokenizedPaymentMethodDTO>(
-                options => options.Type == requestBody.Type && options.Token == requestBody.Token));
-
-        await sutProvider.GetDependency<IStripeAdapter>().Received(1).SubscriptionUpdateAsync(
-            provider.GatewaySubscriptionId, Arg.Is<SubscriptionUpdateOptions>(
-                options => options.CollectionMethod == StripeConstants.CollectionMethod.ChargeAutomatically));
-    }
-
-    #endregion
-
     #region UpdateTaxInformationAsync
 
     [Theory, BitAutoData]
@@ -551,44 +390,4 @@ public class ProviderBillingControllerTests
     }
 
     #endregion
-
-    #region VerifyBankAccount
-
-    [Theory, BitAutoData]
-    public async Task VerifyBankAccount_Ok(
-        Provider provider,
-        VerifyBankAccountRequestBody requestBody,
-        SutProvider<ProviderBillingController> sutProvider)
-    {
-        ConfigureStableAdminInputs(provider, sutProvider);
-
-        var result = await sutProvider.Sut.VerifyBankAccountAsync(provider.Id, requestBody);
-
-        Assert.IsType<Ok>(result);
-
-        await sutProvider.GetDependency<ISubscriberService>().Received(1).VerifyBankAccount(
-            provider,
-            (requestBody.Amount1, requestBody.Amount2));
-    }
-
-    #endregion
-
-    private static void AssertNotFound(IResult result)
-    {
-        Assert.IsType<NotFound<ErrorResponseModel>>(result);
-
-        var response = ((NotFound<ErrorResponseModel>)result).Value;
-
-        Assert.Equal("Resource not found.", response.Message);
-    }
-
-    private static void AssertUnauthorized(IResult result)
-    {
-        Assert.IsType<JsonHttpResult<ErrorResponseModel>>(result);
-
-        var response = (JsonHttpResult<ErrorResponseModel>)result;
-
-        Assert.Equal(StatusCodes.Status401Unauthorized, response.StatusCode);
-        Assert.Equal("Unauthorized.", response.Value.Message);
-    }
 }
